[libpas] Enable libpas on watchOS https://bugs.webkit.org/show_bug.cgi?id=248140 rdar://102556126
Reviewed by NOBODY (OOPS!).

This patch enables libpas on watchOS. watchOS is ARM64_32, which is ARM64 CPU with 32-bit pointers.
Normally libpas cannot work on 32-bit CPUs because it requires 128-bit atomic operations. But since
ARM64_32 is in fact ARM64, we can use 128-bit atomic operations in ARM64_32, thus we can enable
libpas on ARM64_32.

1. To catch the ARM64_32 issue, we enabled GCC_WARN_64_TO_32_BIT_CONVERSION warning on ARM64_32.
2. pas_pair was using uintptr_t. But now we use uint64_t to make it explicit that they are 128-bit atomics.
   This also changes pas_segregated_page_emptiness and pas_versioned_field.
3. pas_mutation_count was using uintptr_t, But now it uses uint64_t internally.

* Source/bmalloc/Configurations/Base.xcconfig:
* Source/bmalloc/bmalloc/BPlatform.h:
* Source/bmalloc/libpas/libpas.xcodeproj/project.pbxproj:
* Source/bmalloc/libpas/src/libpas/pas_basic_heap_config_enumerator_data.c:
(pas_basic_heap_config_enumerator_data_add_page_header_table):
* Source/bmalloc/libpas/src/libpas/pas_bitfit_directory.c:
(pas_bitfit_directory_get_use_epoch):
(pas_bitfit_directory_take_last_empty):
* Source/bmalloc/libpas/src/libpas/pas_bitfit_page_inlines.h:
(pas_bitfit_page_allocation_commit_granules_or_reloop):
(pas_bitfit_page_finish_allocation):
(pas_bitfit_page_allocate):
(pas_bitfit_page_deallocate_with_page_impl):
* Source/bmalloc/libpas/src/libpas/pas_bitfit_size_class.c:
(pas_bitfit_size_class_get_first_free_view):
* Source/bmalloc/libpas/src/libpas/pas_coalign.c:
(pas_coalign_one_sided):
* Source/bmalloc/libpas/src/libpas/pas_compact_tagged_ptr.h:
* Source/bmalloc/libpas/src/libpas/pas_config.h:
* Source/bmalloc/libpas/src/libpas/pas_large_sharing_pool.c:
(boot_tree):
* Source/bmalloc/libpas/src/libpas/pas_local_allocator_inlines.h:
(pas_local_allocator_refill_with_known_config):
* Source/bmalloc/libpas/src/libpas/pas_lock_free_read_ptr_ptr_hashtable.c:
(pas_lock_free_read_ptr_ptr_hashtable_set):
* Source/bmalloc/libpas/src/libpas/pas_lock_free_read_ptr_ptr_hashtable.h:
(pas_lock_free_read_ptr_ptr_hashtable_find):
* Source/bmalloc/libpas/src/libpas/pas_mutation_count.h:
(pas_mutation_count_depend):
* Source/bmalloc/libpas/src/libpas/pas_page_sharing_pool.c:
(get_current_participant):
* Source/bmalloc/libpas/src/libpas/pas_root.h:
* Source/bmalloc/libpas/src/libpas/pas_segregated_directory.c:
(pas_segregated_directory_get_use_epoch):
* Source/bmalloc/libpas/src/libpas/pas_segregated_directory.h:
(pas_segregated_directory_get_last_empty_plus_one_value):
* Source/bmalloc/libpas/src/libpas/pas_segregated_directory_inlines.h:
(pas_segregated_directory_iterate_forward_to_take_first_eligible):
(pas_segregated_directory_iterate_backward_to_take_last_empty):
* Source/bmalloc/libpas/src/libpas/pas_segregated_heap.c:
(pas_segregated_heap_ensure_size_directory_for_size):
* Source/bmalloc/libpas/src/libpas/pas_segregated_page.h:
* Source/bmalloc/libpas/src/libpas/pas_segregated_page_inlines.h:
(pas_segregated_page_deallocate_with_page):
* Source/bmalloc/libpas/src/libpas/pas_status_reporter.c:
(dump_arrow):
(report_bitfit_directory_contents):
(report_segregated_directory_contents):
* Source/bmalloc/libpas/src/libpas/pas_utils.h:
(pas_compare_and_swap_uintptr_weak):
(pas_compare_and_swap_uintptr_strong):
(pas_compare_and_swap_ptr_weak):
(pas_compare_and_swap_ptr_strong):
(pas_pair_create):
(pas_pair_low):
(pas_pair_high):
(pas_compare_and_swap_pair_weak):
(pas_compare_and_swap_pair_strong):
(pas_atomic_store_pair):
(pas_compare_64_opaque):
(pas_compare_ptr_opaque): Deleted.
* Source/bmalloc/libpas/src/libpas/pas_utils_prefix.h:
(__pas_depend64_impl):
(__pas_depend64):
(__pas_depend64_cpu_only):
* Source/bmalloc/libpas/src/libpas/pas_versioned_field.c:
(pas_versioned_field_minimize):
(pas_versioned_field_maximize):
(pas_versioned_field_minimize_watched):
(pas_versioned_field_maximize_watched):
* Source/bmalloc/libpas/src/libpas/pas_versioned_field.h:
(pas_versioned_field_construct):
(pas_versioned_field_create):
(pas_versioned_field_create_with_invalid_version):
(pas_versioned_field_with_version):
(pas_versioned_field_with_value):
(pas_versioned_field_read):
(pas_versioned_field_try_write):
(pas_versioned_field_try_write_watched):
(pas_versioned_field_write):
* Source/bmalloc/libpas/src/mbmalloc/mbmalloc_iso_common_primitive.c:
(mbscavenge):
* Source/bmalloc/libpas/src/test/LargeSharingPoolTests.cpp:

+++ Source/bmalloc/libpas/src/libpas/pas_local_allocator_inlines.h
@@ -1006,10 +1006,9 @@ pas_local_allocator_refill_with_known_config(
     if (verbose) {
         pas_log("old_view = %p.\n", old_view);
         if (old_view) {
-            pas_log("    index = %zu, first_eligible = %zu.\n",
+            pas_log("    index = %zu, first_eligible = %llu.\n",
                     pas_segregated_view_get_index(old_view),
-                    pas_segregated_directory_get_first_eligible(
-                        &pas_segregated_view_get_size_directory(old_view)->base).value);
+                    (unsigned long long)pas_segregated_directory_get_first_eligible(&pas_segregated_view_get_size_directory(old_view)->base).value);
         }
     }
