+++ Source/bmalloc/libpas/src/libpas/pas_segregated_directory.h
@@ -493,11 +493,10 @@ static inline pas_versioned_field pas_segregated_directory_get_last_empty_plus_o
         directory, pas_versioned_field_read);
 }
 
-static inline uintptr_t pas_segregated_directory_get_last_empty_plus_one_value(
+static inline size_t pas_segregated_directory_get_last_empty_plus_one_value(
     pas_segregated_directory* directory)
 {
-    return pas_segregated_directory_get_last_empty_plus_one_impl(
-        directory, pas_versioned_field_read_torn).value;
+    return (size_t)pas_segregated_directory_get_last_empty_plus_one_impl(directory, pas_versioned_field_read_torn).value;
 }
 
 static inline pas_versioned_field pas_segregated_directory_watch_last_empty_plus_one(
